refactor: move pagination helpers to shared module
