+++ StarNews/StarNews/Features/ArticlesList/Sources/ArticlesList/Data/Mappers/ArticlesMapper.swift
@@ -5,6 +5,7 @@
 //  Created by Baher Tamer on 19/06/2025.
 //
 
+import Shared
 import SNCore
 import Foundation.NSData
 
@@ -63,29 +64,5 @@ extension ArticlesMapper {
         let url: String?
         let imageUrl: String?
         let newsSite: String?
-    }
-}
-
-// MARK: - Private Helpers
-extension ArticlesMapper {
-    private func extractLimitAndOffset(from url: String?) -> (limit: Int?, offset: Int?)? {
-        guard
-            let url,
-            let components = URLComponents(string: url),
-            let queryItems = components.queryItems
-        else { return nil }
-        let limit = extractIntQueryItem(named: "limit", from: queryItems)
-        let offset = extractIntQueryItem(named: "offset", from: queryItems)
-        return (limit, offset)
-    }
-    
-    private func extractIntQueryItem(
-        named name: String,
-        from queryItems: [URLQueryItem]
-    ) -> Int? {
-        queryItems
-            .first(where: { $0.name == name })
-            .flatMap { $0.value }
-            .flatMap(Int.init)
     }
 }
