+++ StarNews/StarNews/Features/Shared/Sources/Shared/Domain/Models/Pagination/PaginatedData.swift
