+++ StarNews/StarNews/Features/Shared/Sources/Shared/Presentation/Components/AppLabel.swift
